Introduced App type to simplify startup.
